Only update config.characters with a valid charset
If the set charset contains illegal characters, unavailable in the font, a warning is shown, printing the illegal characters to the user. Also, the characterset won't be updated in this case.

Would it be better to throw an actual error that breaks the sketch and forces the user to fix?