New preEffectFramebuffer initialization
Instead of being an independent framebuffer that draws the canvas as an image to itself, it now receives a reference to the _renderes framebuffer.

That way it's now possible to make use of depth information within the framebuffer.

Also, this should fix a bug related to apple devices, where it thows an error in the console even though the sketch runs fine.