Refactored pre effect framebuffer
It's an independent framebuffer again with the canvas being applied as an image. The depth feature gained didn't really seem  promising in context with ASCII conversion.

With these changes in place, noLoop() now also works properly.

A black canvas would arise if the user provided an empty draw() {} function for some weird reason.