Rewrote sample shader to WebGL1
We'll need to have a function somewhere at some point, which receives certain parameter and returns shader code with those constants set.

When certain actions are done by the user, we then just recreate the shader during run-time to adjust based on the changes made.

Changes to the font size, and thus the grid cell size, would mean that we need to recreate this shader for example, to adjust the SAMPLES_PER_ROW and SAMPLES_PER_COL constants.