Refactored framebuffer initialization.
All framebuffers used in p5.asciify now have the following properties assigned to them. Those help with the sharpness of the overall ascii texture.

antialias: false

depthFormat: UNSIGNED_INT
We are not really dependent on this anyway, also it get's rid of a notification in the console on devices that do not support WebGL2.

textureFiltering: NEAREST